refactoring: move create_disk out and create a voxel_map class to wrap interface with motions
