+++ projects/real_world_ovmm/build_3d_map.py
@@ -282,7 +282,7 @@ def main(
         )
     elif mode == "dir":
         # This is for backwards compatibility with the GOAT data
-        print("WARNING! This is not fully supported yet.")
+        raise NotImplementedError("Output in directory mode not yet working.")
         input_path = Path(input_path)
         pkl_files = input_path.glob("*.pkl")
         sorted_pkl_files = sorted(pkl_files, key=lambda f: int(f.stem))
@@ -299,7 +299,6 @@ def main(
             xyt = np.array([obs.gps[0], obs.gps[1], obs.compass[0]])
             base_pose_matrix = xyt2sophus(xyt).matrix()
             if obs.xyz is None:
-                NotImplementedError("Not yet working")
                 # need to find camera matrix K
                 assert obs.depth is not None, "need depth"
                 xyz = camera.depth_to_xyz(obs.depth)
